requirements.txt
Django<1.10
pillow
django-blog-zinnia>=0.15.1
admin-tools-zinnia>=0.1
django-admin-tools>=0.5.2
django-tinymce==2.2.0
zinnia-wysiwyg-tinymce
django-dbbackup
django-request
django-super-favicon
django-phonenumber-field
django-countries
django-paypal
django-formtools
